refactor: move composed tests outside layers
