+++ extension/test/semantic-diff/github-source-merge.test.ts
@@ -1,14 +1,14 @@
 /// <reference types="node" />
 import { readFileSync } from "node:fs";
 import { beforeAll, describe, expect, it } from "vitest";
-import { BINARY_ASSET } from "../../../test-support/unity-fixtures";
-import type { DiffV2 } from "../../domain/diff/types";
-import type { GuidRepository } from "../../domain/guid/guid-repository";
-import { GithubClient } from "../../infrastructure/clients/github-client";
-import { createDiffer } from "../../infrastructure/clients/wasm-differ-client";
-import { createDiffSession, type DiffContext } from "../diff/create-diff-session";
-import type { DifferGateway } from "../gateway/differ";
-import { mergeGithubSources } from "./github-source-merge";
+import { BINARY_ASSET } from "../../fixtures/unity";
+import { createDiffSession, type DiffContext } from "../../src/application/diff/create-diff-session";
+import type { DifferGateway } from "../../src/application/gateway/differ";
+import { mergeGithubSources } from "../../src/application/internal/github-source-merge";
+import type { DiffV2 } from "../../src/domain/diff/types";
+import type { GuidRepository } from "../../src/domain/guid/guid-repository";
+import { GithubClient } from "../../src/infrastructure/clients/github-client";
+import { createDiffer } from "../../src/infrastructure/clients/wasm-differ-client";
 
 const API_BASE = "https://api.github.test";
 const OWNER = "o";
@@ -93,7 +93,7 @@ function cached(paths: Record<string, string>): MemoryGuidRepository {
 let differ: DifferGateway;
 
 beforeAll(async () => {
-  const bytes = readFileSync(new URL("../../../../zig-out/bin/prefablens.wasm", import.meta.url));
+  const bytes = readFileSync(new URL("../../../zig-out/bin/prefablens.wasm", import.meta.url));
   differ = await createDiffer(bytes);
 });
