+++ extension/test/semantic-diff/prefetch-pr.test.ts
@@ -1,19 +1,19 @@
 /// <reference types="node" />
 import { readFileSync } from "node:fs";
 import { beforeAll, describe, expect, it } from "vitest";
-import { AFTER_PREFAB, BEFORE_PREFAB } from "../../../test-support/unity-fixtures";
-import type { PendingSignIn } from "../../domain/auth/token";
-import type { TokenRepository } from "../../domain/auth/token-repository";
-import type { DiffRepository } from "../../domain/diff/diff-repository";
-import type { DiffV2 } from "../../domain/diff/types";
-import type { GuidMap } from "../../domain/guid/guid-map";
-import type { RepoGuidIndex } from "../../domain/guid/repo-guid-index";
-import type { RepoIndexRepository } from "../../domain/guid/repo-index-repository";
-import { GithubClient } from "../../infrastructure/clients/github-client";
-import { createDiffer } from "../../infrastructure/clients/wasm-differ-client";
-import type { DifferGateway } from "../gateway/differ";
-import { createDiffSession } from "./create-diff-session";
-import { prefetchPr } from "./prefetch-pr";
+import { AFTER_PREFAB, BEFORE_PREFAB } from "../../fixtures/unity";
+import { createDiffSession } from "../../src/application/diff/create-diff-session";
+import { prefetchPr } from "../../src/application/diff/prefetch-pr";
+import type { DifferGateway } from "../../src/application/gateway/differ";
+import type { PendingSignIn } from "../../src/domain/auth/token";
+import type { TokenRepository } from "../../src/domain/auth/token-repository";
+import type { DiffRepository } from "../../src/domain/diff/diff-repository";
+import type { DiffV2 } from "../../src/domain/diff/types";
+import type { GuidMap } from "../../src/domain/guid/guid-map";
+import type { RepoGuidIndex } from "../../src/domain/guid/repo-guid-index";
+import type { RepoIndexRepository } from "../../src/domain/guid/repo-index-repository";
+import { GithubClient } from "../../src/infrastructure/clients/github-client";
+import { createDiffer } from "../../src/infrastructure/clients/wasm-differ-client";
 
 const REQUEST = {
   type: "prefetch" as const,
@@ -76,7 +76,7 @@ class MemoryRepoIndexRepository implements RepoIndexRepository {
 let differ: DifferGateway;
 
 beforeAll(async () => {
-  const bytes = readFileSync(new URL("../../../../zig-out/bin/prefablens.wasm", import.meta.url));
+  const bytes = readFileSync(new URL("../../../zig-out/bin/prefablens.wasm", import.meta.url));
   differ = await createDiffer(bytes);
 });
