+++ extension/tsconfig.json
@@ -14,6 +14,8 @@
   "include": [
     "*.ts",
     "src/**/*",
+    "test/**/*",
+    "fixtures/**/*",
     "e2e/**/*"
   ]
 }
